Added pure Lua 5.1 implementation of RadioGatun[32]
Exclusive or and or-not are implemented with 4x4 S-boxes
(Which is how crypto was done before cache timing attacks made
 that run out of fashion); we also have unused and + or S-boxes

Rotation is done with multiply/add/modulo

Iota step (xor by 1) is one with if/else addition/subtraction

The core beltMill generates a correct RadioGatun[32] test
vector for the string "1234".  I just need to add the
input mapper.