tableStuff: Make API simpler
I have made the API to sPairs simpler: We take a function which
converts a table in to an array with elements in the order we
wish to traverse the table.

This way, we don’t have to understand function closures to make a sorter
which sorts by the values inside a table.

As an aside, this works:

  function closureTest()
    local gg ={foo = 1, bar = 2, baz = 3}
    local ff = function(x, y) return tonumber(gg[x]) > tonumber(gg[y]) end
    local hh = sortedTableKeys(gg, ff)
    for k,v in ipairs(hh) do print (k,v) end
  end

It works because the closure for ff is the local values at the place the
function is created (so ff knows what gg is, even though gg is local to
closureTest() )