wornOutPinPad: Show that 3 unique digits for a 4-digit PIN is most secure
Here’s the deal.  Let us suppose we have a PIN pad, but the numbers are
worn out so the robber knows which digits are in th PIN.  Let us suppose
the PIN is 4 digits long and that robber knows that.  Is it more secure
to have three or four digits in the PIN.

It’s more secure to have three digits (36 combinations) than four digits
(24 combinations).

This Lua script proves that, by brute forcing all 256 of the (1,2,3,4)
possible PINs and discarding ones where the highest digit is higher
than the number of unique digits the PIN has.

Credit: https://idiomdrottning.org/notice/A2nOhYqxMMclHtK6xk