blackCastle: Optional null value argument; handle \ in strings
To correctly add support for \ in JSON strings, this library
would need to have an entire Unicode subsystem: Convert Unicode
codepoints to UTF-8; convert UTF-8 sequences in to Unicode
code points (or, at least, make sure we don’t have invalid UTF-8
in our strings); convert surrogate pairs in to single code points;
etc.  It would make this quick and dirty JSON library about four
times larger.

So, instead, we just pass the string the JSON gives us as-is as
a "binary blob" to Lua.  The UTF-8 infinity symbol `∞` will become
a literal UTF-8 infinity symbol (Hex: E2 88 9E); the sequence
'\u221e' in a JSON string (which represents the infinity symbol)
will remain the ASCII string "\u221e" with the backslash passes as-is.

If blackCastle receives a null keyword, it will, by default convert the
null in the the string `--NULL--`.  This value can be changed by giving
blackCastle an optional second argument.

Also, fully document blackCastle