rg32wiki.lua: Remove categories
Some piece of Mediawiki software out there adds a hidden category
if it doesn’t like the background color of something.  Unfortunately,
this hidden category screws up the RG32 sum of an item since the
category can change the RadioGatun32 input.

Took me too long to debug this.  The workaround is to use a regex
to strip out categories from RadioGatun32 input.